Started identity service implementation
